+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.8
 
 - Fix verbiage for summary/description in scaffolding custom shell gems (change custom widget reference to custom shell)
+- Scaffolding builds a proper binary for custom shell gems and includes it in executables in Rakefile
 
 ## 0.7.7